requirements.txt
flask==3.1.3
flask-cors==6.0.2
pdfplumber==0.11.9
openai==2.26.0
python-dotenv==1.2.2
pytest==9.0.2
pytest-mock==3.15.1
